% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A Hello Function}
\usage{
hello(name = "world", greeting = "Have a great day!")
}
\arguments{
\item{name}{A character string with the name to greet.}

\item{greeting}{A character string with a greeting.}
}
\value{
A character string containing the greeting.
}
\description{
This function prints a friendly greeting.
}
\examples{
hello("Tengis")
}
